Clarify the supportablity of Azure AD only joined clients for Co-management
The intention of these sections, as I understand it, is to call out what is “additional”, for “most” environments – we break it down into 2 main scenarios – starting from SCCM and starting from Intune. 

For those starting with the ConfigMgr client, the “additional” requirement would most likely be Hybrid AzureAD (since they are presumably AD joined). 

For those starting without the ConfigMgr client (i.e. from Intune) and wanting to install the SCCM client from the Internet, we call out the CMG as the additional requirement (in this latter case, Hybrid AD is not noted as a requirement (and it isn’t)).